pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.3.5",
    "datasets>=4.4.1",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.0",
    "i>=2.1.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jq>=1.10.0",
    "jupyter==1.1.1",
    "jupyter-console==6.6.3",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.3.0",
    "jupyter-server==2.17.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.5.0",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.28.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "lark==1.3.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "mistune==3.1.4",
    "msoffcrypto-tool>=5.4.2",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "networkx>=3.6",
    "notebook==7.5.0",
    "notebook-shim==0.2.4",
    "openai>=2.8.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandocfilters==1.5.1",
    "pdfminer>=20191125",
    "pip==25.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.3.9",
    "rank-bm25>=0.2.2",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "umap-learn>=0.5.9.post2",
    "unstructured>=0.18.21",
    "wikipedia>=1.4.0",
]
